requirements.txt
cairosvg==2.8.2
chess==1.11.2
local_wake==0.1.0
numpy==2.3.5
opencv_python==4.12.0.88
Pillow==12.0.0
sounddevice==0.5.3
vosk==0.3.45
